refactor(kernel): split oversized modules (telephony, bt_audio, heorte, sms)
Pure refactor extracting logically independent concerns from four oversized
kernel modules into sub-modules, with re-exports preserving all existing
import paths.

- telephony.rs (1426 -> 1064): parsers -> telephony_parser.rs (298),
  mock transport -> telephony_mock.rs (92)
- bt_audio.rs (1302 -> 674): SBC codec -> sbc.rs (451),
  AVDTP messages -> avdtp.rs (230)
- heorte.rs (1143 -> 546): alarms -> heorte_alarm.rs (220),
  timer/stopwatch -> heorte_timer.rs (435)
- sms.rs (994 -> 733): GSM-7 codec -> gsm7.rs (277)
