+++ crates/thumos/src/telephony.rs
@@ -20,6 +20,11 @@
 //!
 //! Boot integration deferred to Wave 3 (telephony + UI merge via `kinit.rs`).
 //! For now, modules are independently testable via mock `ModemTransport`.
+//!
+//! ## Module structure
+//!
+//! AT response parsers are in [`crate::telephony_parser`].
+//! Mock transport for testing is in [`crate::telephony_mock`].
 
 // WHY: hardware driver API not yet wired to upper layers (Wave 3 integration).
 #![expect(
@@ -27,28 +32,25 @@
     reason = "Telephony driver API not yet wired to kinit (Wave 3)"
 )]
 
-extern crate alloc;
+// Re-export parser functions so external callers can still use crate::telephony::*.
+pub(crate) use crate::telephony_parser::*;
 #[cfg(test)]
-use alloc::vec::Vec;
+pub(crate) use crate::telephony_mock::MockModemTransport;
+
+extern crate alloc;
 
 // ---------------------------------------------------------------------------
 // Constants
 // ---------------------------------------------------------------------------
 
 /// Maximum phone number length in bytes.
-const MAX_NUMBER_LEN: usize = 32;
-
-/// Maximum operator name length in bytes.
-const MAX_OPERATOR_LEN: usize = 32;
+pub(crate) const MAX_NUMBER_LEN: usize = 32;
 
 /// Maximum AT response line length in bytes.
 pub(crate) const MAX_LINE_LEN: usize = 256;
 
 /// Signal strength polling interval in milliseconds (30 seconds).
 const SIGNAL_POLL_INTERVAL_MS: u64 = 30_000;
-
-/// Maximum initialization retry count before entering error state.
-const MAX_INIT_RETRIES: u8 = 3;
 
 // ---------------------------------------------------------------------------
 // Error types
@@ -203,224 +205,6 @@ pub enum TelephonyEvent {
     ModemReady,
     /// Modem entered error state.
     ModemError(TelephonyError),
-}
-
-// ---------------------------------------------------------------------------
-// AT response parsing (no_std, no nom)
-// ---------------------------------------------------------------------------
-
-/// Parse a final result code from an AT response line.
-///
-/// Handles: "OK", "ERROR", "+CME ERROR: <code>"
-pub(crate) fn parse_final_result(line: &[u8]) -> Option<AtResponse> {
-    if line == b"OK" {
-        return Some(AtResponse::Ok);
-    }
-    if line == b"ERROR" {
-        return Some(AtResponse::Error);
-    }
-    if let Some(rest) = strip_prefix(line, b"+CME ERROR: ") {
-        if let Some(code) = parse_u32(rest) {
-            return Some(AtResponse::CmeError(code));
-        }
-    }
-    None
-}
-
-/// Parse a +CSQ response line: "+CSQ: <rssi>,<ber>"
-///
-/// Returns (rssi_raw, ber) where rssi_raw is 0-31 or 99 (unknown).
-pub(crate) fn parse_csq_response(line: &[u8]) -> Option<(u8, u8)> {
-    let rest = strip_prefix(line, b"+CSQ: ")?;
-    let comma = memchr(b',', rest)?;
-    let rssi = parse_u8(&rest[..comma])?;
-    let ber = parse_u8(&rest[comma + 1..])?;
-    Some((rssi, ber))
-}
-
-/// Parse a +CREG response/URC line: "+CREG: <stat>[,<lac>,<ci>]"
-///
-/// We only extract the stat field for telephony purposes.
-fn parse_creg_response(line: &[u8]) -> Option<RegStatus> {
-    let rest = strip_prefix(line, b"+CREG: ")?;
-    // stat is the first field, possibly followed by comma and more fields.
-    let end = memchr(b',', rest).unwrap_or(rest.len());
-    let stat = parse_u8(&rest[..end])?;
-    Some(RegStatus::from(stat))
-}
-
-/// Parse a +COPS? response line: "+COPS: <mode>,<format>,\"<operator>\""
-///
-/// Extracts the operator name string.
-pub(crate) fn parse_cops_response(line: &[u8], name_buf: &mut [u8; MAX_OPERATOR_LEN]) -> Option<u8> {
-    let rest = strip_prefix(line, b"+COPS: ")?;
-    // Find the quoted operator name.
-    let quote_start = memchr(b'"', rest)?;
-    let after_quote = &rest[quote_start + 1..];
-    let quote_end = memchr(b'"', after_quote)?;
-    let name = &after_quote[..quote_end];
-    let len = name.len().min(MAX_OPERATOR_LEN);
-    name_buf[..len].copy_from_slice(&name[..len]);
-    Some(len as u8)
-}
-
-/// Parse a +CLIP URC line: "+CLIP: \"<number>\",<type>..."
-///
-/// Extracts the caller phone number.
-fn parse_clip_response(line: &[u8], number_buf: &mut [u8; MAX_NUMBER_LEN]) -> Option<u8> {
-    let rest = strip_prefix(line, b"+CLIP: ")?;
-    // Number is in quotes.
-    let quote_start = memchr(b'"', rest)?;
-    let after_quote = &rest[quote_start + 1..];
-    let quote_end = memchr(b'"', after_quote)?;
-    let number = &after_quote[..quote_end];
-    let len = number.len().min(MAX_NUMBER_LEN);
-    number_buf[..len].copy_from_slice(&number[..len]);
-    Some(len as u8)
-}
-
-/// Parse a +CPIN? response line: "+CPIN: <status>"
-///
-/// Returns true if the SIM is ready (no PIN required).
-fn parse_cpin_response(line: &[u8]) -> Option<bool> {
-    let rest = strip_prefix(line, b"+CPIN: ")?;
-    if rest == b"READY" {
-        Some(true)
-    } else if starts_with(rest, b"SIM PIN") {
-        Some(false)
-    } else {
-        // Other states (SIM PUK, etc.) — treat as not ready.
-        Some(false)
-    }
-}
-
-/// Check if a line is a RING URC.
-fn is_ring(line: &[u8]) -> bool {
-    line == b"RING"
-}
-
-/// Check if a line is a NO CARRIER URC.
-fn is_no_carrier(line: &[u8]) -> bool {
-    line == b"NO CARRIER"
-}
-
-/// Check if a line is a BUSY URC.
-fn is_busy(line: &[u8]) -> bool {
-    line == b"BUSY"
-}
-
-/// Try to parse a line as any known URC.
-fn parse_urc(line: &[u8]) -> Option<Urc> {
-    if is_ring(line) {
-        return Some(Urc::Ring);
-    }
-    if is_no_carrier(line) {
-        return Some(Urc::NoCarrier);
-    }
-    if is_busy(line) {
-        return Some(Urc::Busy);
-    }
-    if let Some((rssi, ber)) = parse_csq_response(line) {
-        return Some(Urc::Csq { rssi, ber });
-    }
-    if let Some(stat) = parse_creg_response(line) {
-        return Some(Urc::Creg { stat });
-    }
-    if starts_with(line, b"+CLIP: ") {
-        let mut number = [0u8; MAX_NUMBER_LEN];
-        if let Some(len) = parse_clip_response(line, &mut number) {
-            return Some(Urc::Clip {
-                number,
-                number_len: len,
-            });
-        }
-    }
-    None
-}
-
-// ---------------------------------------------------------------------------
-// Byte-level parsing helpers (no_std, no alloc)
-// ---------------------------------------------------------------------------
-
-/// Strip a prefix from a byte slice. Returns `None` if the prefix doesn't match.
-fn strip_prefix<'a>(input: &'a [u8], prefix: &[u8]) -> Option<&'a [u8]> {
-    if input.len() >= prefix.len() && &input[..prefix.len()] == prefix {
-        Some(&input[prefix.len()..])
-    } else {
-        None
-    }
-}
-
-/// Check if `input` starts with `prefix`.
-fn starts_with(input: &[u8], prefix: &[u8]) -> bool {
-    input.len() >= prefix.len() && &input[..prefix.len()] == prefix
-}
-
-/// Find the position of a byte in a slice.
-fn memchr(needle: u8, haystack: &[u8]) -> Option<usize> {
-    haystack.iter().position(|&b| b == needle)
-}
-
-/// Parse a byte slice as a decimal u8.
-fn parse_u8(input: &[u8]) -> Option<u8> {
-    if input.is_empty() {
-        return None;
-    }
-    let mut result: u8 = 0;
-    for &b in input {
-        if !b.is_ascii_digit() {
-            return None;
-        }
-        result = result.checked_mul(10)?.checked_add(b - b'0')?;
-    }
-    Some(result)
-}
-
-/// Parse a byte slice as a decimal u32.
-fn parse_u32(input: &[u8]) -> Option<u32> {
-    if input.is_empty() {
-        return None;
-    }
-    let mut result: u32 = 0;
-    for &b in input {
-        if !b.is_ascii_digit() {
-            return None;
-        }
-        result = result.checked_mul(10)?.checked_add(u32::from(b - b'0'))?;
-    }
-    Some(result)
-}
-
-/// Convert raw AT+CSQ RSSI value (0-31, 99=unknown) to dBm.
-///
-/// Formula: dBm = -113 + (rssi * 2), per 3GPP TS 27.007.
-fn rssi_to_dbm(rssi: u8) -> i16 {
-    if rssi == 99 {
-        return -999; // unknown sentinel
-    }
-    -113 + (i16::from(rssi) * 2)
-}
-
-/// Map signal strength in dBm to bars (0-4).
-///
-/// Thresholds based on the spec's dBm mapping:
-/// - >= -70 dBm -> 4 bars (excellent)
-/// - >= -85 dBm -> 3 bars (good)
-/// - >= -100 dBm -> 2 bars (fair)
-/// - >= -110 dBm -> 1 bar  (poor)
-/// - <  -110 dBm -> 0 bars (no signal)
-pub fn dbm_to_bars(dbm: i16) -> u8 {
-    if dbm >= -70 {
-        4
-    } else if dbm >= -85 {
-        3
-    } else if dbm >= -100 {
-        2
-    } else if dbm >= -110 {
-        1
-    } else {
-        0
-    }
 }
 
 // ---------------------------------------------------------------------------
@@ -1077,96 +861,6 @@ impl ModemTransport for CcciModemTransport {
 }
 
 // ---------------------------------------------------------------------------
-// Mock transport (test only)
-// ---------------------------------------------------------------------------
-
-/// Mock modem transport for unit testing.
-///
-/// Records sent commands and replays pre-configured responses.
-#[cfg(test)]
-pub struct MockModemTransport {
-    /// AT commands sent via `send_at`.
-    pub sent_commands: Vec<Vec<u8>>,
-    /// Response lines to return from `recv_line`, in FIFO order.
-    pub response_lines: Vec<Vec<u8>>,
-    /// URC lines to return from `poll_urc_line`, in FIFO order.
-    pub urc_lines: Vec<Vec<u8>>,
-    /// Whether send_at should succeed.
-    pub send_ok: bool,
-}
-
-#[cfg(test)]
-impl MockModemTransport {
-    /// Create a new mock transport with all operations succeeding.
-    pub fn new() -> Self {
-        Self {
-            sent_commands: Vec::new(),
-            response_lines: Vec::new(),
-            urc_lines: Vec::new(),
-            send_ok: true,
-        }
-    }
-
-    /// Queue a response line to be returned by `recv_line`.
-    pub fn queue_response(&mut self, line: &[u8]) {
-        self.response_lines.push(line.to_vec());
-    }
-
-    /// Queue a URC line to be returned by `poll_urc_line`.
-    pub fn queue_urc(&mut self, line: &[u8]) {
-        self.urc_lines.push(line.to_vec());
-    }
-
-    /// Queue a simple "OK" response.
-    pub fn queue_ok(&mut self) {
-        self.queue_response(b"OK");
-    }
-
-    /// Queue an info line followed by "OK".
-    pub fn queue_info_ok(&mut self, info: &[u8]) {
-        self.queue_response(info);
-        self.queue_response(b"OK");
-    }
-}
-
-#[cfg(test)]
-impl ModemTransport for MockModemTransport {
-    fn send_at(&mut self, command: &str) -> Result<(), TelephonyError> {
-        if !self.send_ok {
-            return Err(TelephonyError::TransportError);
-        }
-        self.sent_commands.push(command.as_bytes().to_vec());
-        Ok(())
-    }
-
-    fn recv_line(
-        &mut self,
-        buf: &mut [u8; MAX_LINE_LEN],
-        _timeout_ms: u32,
-    ) -> Result<usize, TelephonyError> {
-        if let Some(line) = self.response_lines.first() {
-            let len = line.len().min(MAX_LINE_LEN);
-            buf[..len].copy_from_slice(&line[..len]);
-            self.response_lines.remove(0);
-            Ok(len)
-        } else {
-            Err(TelephonyError::Timeout)
-        }
-    }
-
-    fn poll_urc_line(&mut self, buf: &mut [u8; MAX_LINE_LEN]) -> Option<usize> {
-        if let Some(line) = self.urc_lines.first() {
-            let len = line.len().min(MAX_LINE_LEN);
-            buf[..len].copy_from_slice(&line[..len]);
-            self.urc_lines.remove(0);
-            Some(len)
-        } else {
-            None
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
 
@@ -1356,62 +1050,6 @@ mod tests {
             *tel.call_state(),
             CallState::Idle,
             "call state must transition to Idle on NO CARRIER"
-        );
-    }
-
-    #[test]
-    fn signal_strength_maps_to_correct_bars() {
-        // Test the dBm-to-bars mapping at boundary values.
-        assert_eq!(dbm_to_bars(-70), 4, "-70 dBm must be 4 bars");
-        assert_eq!(dbm_to_bars(-69), 4, "-69 dBm must be 4 bars");
-        assert_eq!(dbm_to_bars(-71), 3, "-71 dBm must be 3 bars");
-        assert_eq!(dbm_to_bars(-85), 3, "-85 dBm must be 3 bars");
-        assert_eq!(dbm_to_bars(-86), 2, "-86 dBm must be 2 bars");
-        assert_eq!(dbm_to_bars(-100), 2, "-100 dBm must be 2 bars");
-        assert_eq!(dbm_to_bars(-101), 1, "-101 dBm must be 1 bar");
-        assert_eq!(dbm_to_bars(-110), 1, "-110 dBm must be 1 bar");
-        assert_eq!(dbm_to_bars(-111), 0, "-111 dBm must be 0 bars");
-        assert_eq!(dbm_to_bars(-999), 0, "unknown signal must be 0 bars");
-    }
-
-    #[test]
-    fn parse_csq_response_extracts_rssi() {
-        let line = b"+CSQ: 18,99";
-        let result = parse_csq_response(line);
-        assert_eq!(
-            result,
-            Some((18, 99)),
-            "CSQ response must extract rssi=18 and ber=99"
-        );
-
-        // Verify dBm conversion: RSSI 18 => -113 + (18*2) = -77 dBm.
-        let dbm = rssi_to_dbm(18);
-        assert_eq!(dbm, -77, "RSSI 18 must convert to -77 dBm");
-    }
-
-    #[test]
-    fn parse_cops_response_extracts_operator() {
-        let line = b"+COPS: 0,0,\"T-Mobile\"";
-        let mut name = [0u8; MAX_OPERATOR_LEN];
-        let len = parse_cops_response(line, &mut name);
-        assert_eq!(len, Some(8), "operator name length must be 8");
-        assert_eq!(
-            &name[..8],
-            b"T-Mobile",
-            "operator name must be T-Mobile"
-        );
-    }
-
-    #[test]
-    fn parse_clip_response_extracts_number() {
-        let line = b"+CLIP: \"+15551234567\",145";
-        let mut number = [0u8; MAX_NUMBER_LEN];
-        let len = parse_clip_response(line, &mut number);
-        assert_eq!(len, Some(12), "number length must be 12");
-        assert_eq!(
-            &number[..12],
-            b"+15551234567",
-            "caller ID number must be +15551234567"
         );
     }
